Particle system finalized + clean up
In addition to completed particle system, this update includes all the internal documentation for the project and a guide document for adding new functionality. The serialization process for all the world props has been completed. A small update for water system - shallow remaining water evaporates to allow editor to create objects on the ground and create new blocks in the empty cells.